# openAPI-def.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "doughleap",
    "version": "1.0.0"
  },
  "paths": {
...
    "/tags/recipe": {
      "get": {
        "tags": [
          "Tags"
        ],
        "operationId": "getRecipeTags",
        "parameters": [
          {
            "name": "recipeTagIDs",
            "in": "query",
            "description": "Comma-separated list of recipe tag IDs.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer"
              }
            },
            "style": "form",
            "explode": false
          },
          {
            "name": "recipeID",
            "in": "query",
            "description": "The ID of the recipe.",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "tagID",
            "in": "query",
            "description": "The ID of the tag.",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "recipeTagID": {
                        "type": "integer",
                        "description": "The ID of the recipe tag."
                      },
                      "recipeID": {
                        "type": "integer",
                        "description": "The ID of the recipe."
                      },
                      "tagID": {
                        "type": "integer",
                        "description": "The ID of the tag."
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tags"
        ],
        "operationId": "createRecipeTag",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "recipeID",
                  "tagID"
                ],
                "properties": {
                  "recipeID": {
                    "type": "integer",
                    "description": "The ID of the recipe."
                  },
                  "tagID": {
                    "type": "integer",
                    "description": "The ID of the tag."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "recipeTagID": {
                      "type": "integer",
                      "description": "The ID of the recipe tag."
                    },
                    "recipeID": {
                      "type": "integer",
                      "description": "The ID of the recipe."
                    },
                    "tagID": {
                      "type": "integer",
                      "description": "The ID of the tag."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tags/recipe/{recipeTagID}": {
      "delete": {
        "tags": [
          "Tags"
        ],
        "operationId": "deleteRecipeTag",
        "parameters": [
          {
            "name": "recipeTagID",
            "in": "path",
            "description": "The ID of the recipe tag.",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/ingredients/recipe": {
      "get": {
        "tags": [
          "Ingredients"
        ],
        "operationId": "getRecipeIngredients",
        "parameters": [
          {
            "name": "recipeIngredientIDs",
            "in": "query",
            "description": "Comma-separated list of recipe ingredient IDs.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer"
              }
            },
            "style": "form",
            "explode": false
          },
          {
            "name": "recipeID",
            "in": "query",
            "description": "The ID of the recipe.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          },
          {
            "name": "ingredientID",
            "in": "query",
            "description": "The ID of the ingredient.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "recipeIngredientID": {
                        "type": "integer",
                        "description": "The ID of the recipe-ingredient pairing."
                      },
                      "recipeID": {
                        "type": "integer",
                        "description": "The ID of the recipe."
                      },
                      "ingredientID": {
                        "type": "integer",
                        "description": "The ID of the ingredient."
                      },
                      "measurementUnit": {
                        "type": "string",
                        "description": "The measurement unit for the ingredient."
                      },
                      "measurement": {
                        "type": "integer",
                        "description": "The measurement value for the ingredient."
                      }
                    },
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Ingredients"
        ],
        "operationId": "createRecipeIngredient",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "recipeID",
                  "ingredientID",
                  "measurementUnit",
                  "measurement"
                ],
                "properties": {
                  "recipeID": {
                    "type": "integer",
                    "description": "The ID of the recipe."
                  },
                  "ingredientID": {
                    "type": "integer",
                    "description": "The ID of the ingredient."
                  },
                  "measurementUnit": {
                    "type": "string",
                    "description": "The measurement unit for the ingredient.",
                    "enum": [
                      "gram",
                      "ounce",
                      "pound",
                      "teaspoon",
                      "tablespoon",
                      "cup",
                      "quart",
                      "gallon",
                      "milliliter",
                      "liter",
                      "bags",
                      "boxes",
                      "cartons",
                      "pallets",
                      "bottles",
                      "containers",
                      "bunch",
                      "dash"
                    ]
                  },
                  "measurement": {
                    "type": "integer",
                    "description": "The measurement value for the ingredient."
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "recipeIngredientID": {
                      "type": "integer",
                      "description": "The ID of the recipe-ingredient pairing."
                    },
                    "recipeID": {
                      "type": "integer",
                      "description": "The ID of the recipe."
                    },
                    "ingredientID": {
                      "type": "integer",
                      "description": "The ID of the ingredient."
                    },
                    "measurementUnit": {
                      "type": "string",
                      "description": "The measurement unit for the ingredient."
                    },
                    "measurement": {
                      "type": "integer",
                      "description": "The measurement value for the ingredient."
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/ingredients/recipe/{recipeIngredientID}": {
      "patch": {
        "tags": [
          "Ingredients"
        ],
        "operationId": "updateRecipeIngredient",
        "parameters": [
          {
            "name": "recipeIngredientID",
            "in": "path",
            "description": "The ID of the recipe ingredient.",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "measurementUnit": {
                    "type": "string",
                    "description": "The updated measurement unit for the recipe ingredient.",
                    "enum": [
                      "gram",
                      "ounce",
                      "pound",
                      "teaspoon",
                      "tablespoon",
                      "cup",
                      "quart",
                      "gallon",
                      "milliliter",
                      "liter",
                      "bags",
                      "boxes",
                      "cartons",
                      "pallets",
                      "bottles",
                      "containers"
                    ]
                  },
                  "measurement": {
                    "type": "integer",
                    "description": "The updated measurement value for the recipe ingredient."
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "recipeIngredientID": {
                      "type": "integer",
                      "description": "The ID of the recipe-ingredient pairing."
                    },
                    "recipeID": {
                      "type": "integer",
                      "description": "The ID of the recipe."
                    },
                    "ingredientID": {
                      "type": "integer",
                      "description": "The ID of the ingredient."
                    },
                    "measurementUnit": {
                      "type": "string",
                      "description": "The measurement unit for the ingredient."
                    },
                    "measurement": {
                      "type": "integer",
                      "description": "The measurement value for the ingredient."
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Ingredients"
        ],
        "operationId": "deleteRecipeIngredient",
        "parameters": [
          {
            "name": "recipeIngredientID",
            "in": "path",
            "description": "The ID of the recipe ingredient.",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  },
...
}